AI Doctor/requirements.txt
flask
Flask-Migrate
requests
Flask-SQLAlchemy